Add files via upload
The goal of this project is to implement a (very) rudimentary database engine that is loosely based on a
hybrid between MySQL and SQLite, which I call DavisBase. The implementation operates entirely from the command line and API calls (no GUI). The indexed search and B+ tree implementation is also added as a Bonus. The boilerplate code was provided by Chris Irwin Davis and the assignment problem statement was also provided from him